+++ test_pytest_flake8dir.py
@@ -5,16 +5,30 @@ import pytest
 import six
 
 
-def test_simple_run(flake8dir):
-    flake8dir.make_py_files(
-        example="""
+def test_make_py_files_single(flake8dir):
+    flake8dir.make_py_files(example="""
         x  = 1
-        """
-    )
+    """)
     result = flake8dir.run_flake8()
     assert result.out_lines == [
         './example.py:1:2: E221 multiple spaces before operator'
     ]
+
+
+def test_make_py_files_double(flake8dir):
+    flake8dir.make_py_files(
+        example1="""
+            x  = 1
+        """,
+        example2="""
+            y  = 2
+        """
+    )
+    result = flake8dir.run_flake8()
+    assert set(result.out_lines) == {
+        './example1.py:1:2: E221 multiple spaces before operator',
+        './example2.py:1:2: E221 multiple spaces before operator',
+    }
 
 
 def test_make_py_files_no_positional_args(flake8dir):
@@ -31,17 +45,17 @@ def test_make_py_files_requires_at_least_one_kwarg(flake8dir):
     assert 'make_py_files requires at least one keyword argument' in six.text_type(excinfo.value)
 
 
-def test_passing_args(flake8dir):
-    flake8dir.make_py_files(
-        example="""
+def test_make_example_py(flake8dir):
+    flake8dir.make_example_py("""
         x  = 1
-        """
-    )
-    result = flake8dir.run_flake8(extra_args=['--ignore', 'E221'])
-    assert result.out_lines == []
+    """)
+    result = flake8dir.run_flake8()
+    assert result.out_lines == [
+        './example.py:1:2: E221 multiple spaces before operator'
+    ]
 
 
-def test_setup_cfg(flake8dir):
+def test_make_setup_cfg(flake8dir):
     flake8dir.make_setup_cfg("""
         [flake8]
         ignore = E221
@@ -52,6 +66,26 @@ def test_setup_cfg(flake8dir):
         """
     )
     result = flake8dir.run_flake8()
+    assert result.out_lines == []
+
+
+def test_make_file(flake8dir):
+    flake8dir.make_file('myexample.py', """
+        x  = 1
+    """)
+    result = flake8dir.run_flake8()
+    assert result.out_lines == [
+        './myexample.py:1:2: E221 multiple spaces before operator'
+    ]
+
+
+def test_extra_args(flake8dir):
+    flake8dir.make_py_files(
+        example="""
+        x  = 1
+        """
+    )
+    result = flake8dir.run_flake8(extra_args=['--ignore', 'E221'])
     assert result.out_lines == []