+++ HISTORY.rst
@@ -8,6 +8,8 @@ Pending Release
 
 .. Insert new release notes below this line
 
+* Add convenience methods ``make_example_py`` and ``make_file``.
+
 1.0.0 (2017-06-22)
 ------------------